pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["hanoi0126 <ipocchi129@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
torch = "^2.5.0"
einops = "^0.8.0"
numpy = "^2.1.2"
matplotlib = "^3.9.2"
scipy = "^1.14.1"
torchvision = "^0.20.0"
wandb = "^0.18.5"
python-dotenv = "^1.0.1"
hydra-core = "^1.3.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.12.1"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
